pyproject.toml
[project]
name = "assessiment-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.8.1",
    "fastapi>=0.116.1",
    "groq>=0.31.0",
    "ipywidgets>=8.1.7",
    "openai>=1.102.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pip>=25.2",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "tavily-python>=0.7.11",
    "uvicorn>=0.35.0",
]
